updated apimanager class, added exception handling, optimized for django backend
