+++ requirements.txt
@@ -7,3 +7,5 @@ psycopg2-binary==2.9.1
 python-dotenv==0.19.1
 pytz==2021.3
 sqlparse==0.4.2
+requests==2.26.0
+alpaca_trade_api==1.4.1
